user create, get, findOne, delete done setup

+++ src/modules/users/service.user.ts
@@ -1,4 +1,5 @@
-import { TUser } from "./interface.user";
+import mongoose from "mongoose";
+import { TOrder, TUser } from "./interface.user";
 import { User } from "./mode.user";
 
 const createUser = async (userData: TUser) => {
@@ -8,48 +9,48 @@ const createUser = async (userData: TUser) => {
   }
   const result = await User.create(userData);
 
-  // # instance
-  // built in instance method
-  // const student = new Student(studentData);
-  // if (await student.isUserExists(studentData.id)) {
-  //   throw new Error('student already exists');
-  // }
-  // const result = await student.save();
-
   return result;
 };
 
 const getAllUser = async () => {
-  const result = await User.find();
+  const result = await User.aggregate([
+    { $project: { userName: 1, fullName: 1, age: 1, email: 1, address: 1 } },
+  ]);
   return result;
 };
 
 // get single user
 const getSingleUser = async (id: string) => {
-  const result = await User.findOne({ id });
+  const result = await User.findOne(
+    { userId: id },
+    { userId: 1, userName: 1, fullName: 1, age: 1, email: 1, address: 1 }
+  );
   return result;
+
+  // const result = User.aggregate([{ $match: { userId: id } }]);
 };
 
 const updateUser = async (id: string, data: TUser) => {
-  const result = await User.updateOne({ id }, {}, data);
+  const result = await User.updateOne(
+    { userId: id },
+    { $set: data },
+    {
+      new: true,
+    }
+  );
+  console.log("up", result);
+  return result;
+};
+const updateUserOrder = async (id: string, data: TOrder) => {
+  const result = await User.addProductToOrders(Number(id), data);
+  // console.log("up", result);
   return result;
 };
 
+// delete user
 const deleteUser = async (id: string) => {
-  if (await User.isUserExists(id)) {
-    const result = await User.deleteOne({ id });
-    return result;
-  }
-  throw new Error(
-    JSON.stringify({
-      success: false,
-      message: "User not found",
-      error: {
-        code: 404,
-        description: "User not found!",
-      },
-    })
-  );
+  const result = await User.deleteOne({ userId: id });
+  return result;
 };
 
 export const UserServices = {
@@ -58,4 +59,5 @@ export const UserServices = {
   getSingleUser,
   updateUser,
   deleteUser,
+  updateUserOrder,
 };
